Add clean() function
This is still a work in progress. The function takes device and an
optional 'config' argument. Device must be a Device instance and
config must be a path to a valid cleaner config file.

The function parses the config file and can either remove_recursive
and remove files or directories, uninstall apps or replace remote
files with local files.

See 'cleaner_config' file for config specifications.