Rmeove helper_interactive module
I have plans for reintroducing it at a later date, but it will be
an actual GUI (probably in QT). This was a relic of the past, when
this project was a bunch of interactive batch scripts (horrible
mess and it didn't always work because of reasons, but the part for
installing apps concurrently on arbitrary number of devices was fun
to write).