Remove print calls with "\n" as only argument
Whenever an extra preceeding empty line is needed for a string
(readability, or any other purpose), it must be included at the
start of that string.

This is done to unify the look of messages in the CLI and GUI
console, since GUI console will now skip messages containing only
whitespace.