Remove device connection checking from main
This removes a tiny bit of complexity from the spaghetti-filled
main module and makes detecting when devices suddenly disconnect
at a top level possible with a simple try-except.

Note: CLI batch commands will have to be adapted to allow
command execution on connected devices if one of them disconnects.
I'm basically falling asleep, so it's entirely possible that I will
revert it as soon as I come to my senses.