New year, new rewrite of data extraction module! #4
While the previous incarnation of this module was a horrible mess,
the ability to perform data extraction lazily is pretty neat.

With this rewrite I decided to go back to implementation that
focuses on getting the data from the given section, and not about
exhausting the data source. The potential for lazy extraction is
greatly limited--although I could just force Device to keep the
output from data source until all functions that could have used
it have done so, but it feels wasteful and just _wrong_ to do so.

The laziness potential could be improved with some planning and
better segmentation of different sections of extracted info, but
I think the focus for now should be making sure that it actually
works across as many devices as possible, and finding out ways
for cutting the amount of data being extracted at once to only
the needed minimum.

Also: this breaks everything!